Move out dumper interface from spi package
